andamios/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.10.0
dj-database-url==0.5.0
Django==3.1.2
django-admin-interface==0.17.3
django-bootstrap5==2.1.2
django-colorfield==0.4.5
django-crispy-forms==1.13.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
humanize==3.12.0
Pillow==8.4.0
python-decouple==3.5
pytz==2021.3
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
whitenoise==5.3.0